// src/app/components/models/evaluation.models.ts
export interface Evaluation {
  uid: string;
  q1: string;
  q2: string;
  q3: string;
  q4: string;
  q5: string;
  q6: string;
  q7: string;
  q8: string;
  q9: string;
  q10: string;
  q11: string;
  q12: string;
  q13: string;
  q14: string;
  q15: string;
  q16: string;
  q17: string;
  q18: string;
  q19: string;
  q20: string;
  q21: string;
  q22: string;
  q23: string;
  q24: string;
  q25: string;
  token: string;
  program_id: string;
  certificate_id: string;
  programs_section_id: string;
  question_id: string;
  number_of_questions: number;
}